Update README.md
Microservice that works between with my partner's main Mad Libs program. The Microservice takes the number of the mad lib story given by the user (which is written into the comm_pipe.txt file) and then populates the mad lib story into the story.txt file. This story and number of mad lib words to be replaced is then received by the main Mad Libs program.